+++ EventLogger/Scenes/Schedule/View/DateRangeFieldContainerView.swift
@@ -357,41 +357,6 @@ final class DateRangeFieldContainerView: UIView {
     }
 }
 
-extension UIButton {
-    /// 버튼들 생김새 통합 설정
-    static func makeDateButton() -> UIButton {
-        let button = UIButton(type: .system)
-        var config = UIButton.Configuration.bordered()
-        config.contentInsets = .init(top: 2, leading: 8, bottom: 2, trailing: 8)
-        config.background.cornerRadius = 6
-        config.baseForegroundColor = .white
-        config.background.backgroundColor = .systemGray3
-        config.titleTextAttributesTransformer = UIConfigurationTextAttributesTransformer { incoming in
-            var outgoing = incoming
-            outgoing.font = .font17Regular
-            return outgoing
-        }
-        button.configuration = config
-        button.configurationUpdateHandler = { button in
-            var config = button.configuration ?? .bordered()
-            let isOn = button.isSelected
-            config.baseForegroundColor = isOn ? .systemOrange : .white
-            button.configuration = config
-        }
-        return button
-    }
-}
-
-extension UIDatePicker {
-    func applyYearRange(minYear: Int, maxYear: Int) {
-        let calendar = Calendar(identifier: .gregorian)
-        var min = DateComponents(); min.year = minYear; min.month = 1; min.day = 1
-        var max = DateComponents(); max.year = maxYear; max.month = 1; max.day = 1
-        minimumDate = calendar.date(from: min)
-        maximumDate = calendar.date(from: max)
-    }
-}
-
 #Preview {
     DateRangeFieldContainerView()
 }
